List > Sorted Set

+++ src/main/java/com/example/yetiproject/dto/ticket/TicketRequestDto.java
@@ -9,7 +9,7 @@ import lombok.*;
 @AllArgsConstructor
 public class TicketRequestDto {
 	private Long userId;
-	private Double now;
+	private Long now;
 	private Long ticketInfoId;
 	private Long posX;
 	private Long posY;
